finish lab

+++ table.c
@@ -45,7 +45,7 @@ TAB_table TAB_empty(void)
 void TAB_enter(TAB_table t, void *key, void *value)
 {int index;
  assert(t && key);
- index = ((unsigned)key) % TABSIZE;
+ index = ((unsigned long)key) % TABSIZE;
  t->table[index] = Binder(key, value,t->table[index], t->top);
  t->top = key;
 }
@@ -54,7 +54,7 @@ void *TAB_look(TAB_table t, void *key)
 {int index;
  binder b;
  assert(t && key);
- index=((unsigned)key) % TABSIZE;
+ index=((unsigned long)key) % TABSIZE;
  for(b=t->table[index]; b; b=b->next)
    if (b->key==key) return b->value;
  return NULL;
@@ -65,7 +65,7 @@ void *TAB_pop(TAB_table t) {
   assert (t);
   k = t->top;
   assert (k);
-  index = ((unsigned)k) % TABSIZE;
+  index = ((unsigned long)k) % TABSIZE;
   b = t->table[index];
   assert(b);
   t->table[index] = b->next;
@@ -75,7 +75,7 @@ void *TAB_pop(TAB_table t) {
 
 void TAB_dump(TAB_table t, void (*show)(void *key, void *value)) {
   void *k = t->top;
-  int index = ((unsigned)k) % TABSIZE;
+  int index = ((unsigned long)k) % TABSIZE;
   binder b = t->table[index];
   if (b==NULL) return;
   t->table[index]=b->next;
